\name{odfPageBreak}
\alias{odfPageBreak}
\title{Generate a Page Break}
\description{
  Generate a page break in an odfWeave document.
}
\usage{
odfPageBreak()
}
\details{
  This function should be called in a code chunk in an odfWeave document
  at the point where a page break is desired.
}
\examples{
\dontrun{
odfPageBreak()
}
}
\keyword{utilities}
