\name{odfTmpDir}
\alias{odfTmpDir}
\title{odfWeave working directory}
\description{
Create a working directory for odfWeave
}
\usage{
odfTmpDir()
}
\details{
This function determines the temp directory and creates a subdirectory. If a subdirectory cannot be created, an error is thrown.
}
\value{
a character string containing the path
}
\author{Max Kuhn}

\keyword{utilities}

