\name{odfWeave-package}
\alias{odfWeave-package}
\docType{package}
\title{
Sweave-like capability for OpenDocument files
}
\description{
This package tries to mimic the functionality of \code{\link{Sweave}} for OpenDocuemnt files (I've only used it for writer documents though). The OpenDocument format (ODF) is very different from latex (for example it is a compressed file) and due to these differences, it wasn't possible to actually use \code{\link{Sweave}}.
}
\details{
\tabular{ll}{
Package: \tab odfWeave\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2006-03-09\cr
License: \tab GPL 2\cr
}

}
\author{
Max Kuhn

Maintainer: Max Kuhn <max.kuhn@pfizer.com>
}
\references{
}
\keyword{ utilities }

