\name{odfTable}
\alias{odfTable}
\title{Generic method for creating a table}
\description{
Generic method for creating a table}
\usage{
odfTable(x, ...)
}
\arguments{
  \item{x}{a vector, matrix or data frame}
  \item{\dots}{optional arguments}
}
\details{
There are methods for vectors, matrices and data frames
}

\author{Max Kuhn}

\keyword{utilities}

