\name{odfWeave-package}
\alias{odfWeave-package}
\docType{package}
\title{
Sweave-like capability for Open Document Format (ODF) files
}
\description{
This package tries to mimic the functionality of \code{\link{Sweave}} for OpenDocuemnt files (I've only used it for writer documents though).}

\details{
\tabular{ll}{
Package: \tab odfWeave\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2006-07-01\cr
License: \tab GPL 2\cr
}

}
\author{
Max Kuhn

Maintainer: Max Kuhn <max.kuhn@pfizer.com>
}
\references{
}
\keyword{ utilities }

